_create_toml_file method
